Cleanup, added coins

+++ Assets/ARDK/AR/Awareness/_AwarenessBufferHelper.cs
@@ -1,15 +1,11 @@
 // Copyright 2022 Niantic, Inc. All Rights Reserved.
 
 using System;
-using System.Collections.Generic;
 using System.Runtime.CompilerServices;
-
 using Niantic.ARDK.Utilities;
 using Niantic.ARDK.Utilities.Logging;
-
 using Unity.Collections;
 using Unity.Collections.LowLevel.Unsafe;
-
 using UnityEngine;
 
 namespace Niantic.ARDK.AR.Awareness
@@ -106,8 +102,8 @@ namespace Niantic.ARDK.AR.Awareness
 
       int cropStartX = 0, cropStartY = 0;
 
-      newWidth = (int)srcWidth;
-      newHeight = (int)srcHeight;
+      newWidth = srcWidth;
+      newHeight = srcHeight;
 
       if (srcRatio > trgRatio)
       {
